+++ util/opentelemetry-util-genai/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change `InferenceInvocation` init params to only accept base params
 - Pass in `attributes` on invocation `_start` so samplers have access to attributes.
   ([#4538](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4538))
+- Apply attribute for sampling on instantiation of all invocation types.
+  ([#4553](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4553))
 
 ## Version 0.4b0 (2026-05-01)